Code refactor to split logic for DeltaSharedTableLoader and DeltaSharedTable
