+++ server/src/main/scala/io/delta/standalone/internal/DeltaSharedTable.scala
@@ -25,7 +25,6 @@ import java.util.concurrent.TimeUnit
 import scala.collection.JavaConverters._
 
 import com.google.cloud.hadoop.fs.gcs.GoogleHadoopFileSystem
-import com.google.common.cache.CacheBuilder
 import com.google.common.hash.Hashing
 import io.delta.standalone.DeltaLog
 import io.delta.standalone.internal.actions.{AddCDCFile, AddFile, Metadata, Protocol, RemoveFile}
@@ -43,50 +42,9 @@ import scala.util.control.NonFatal
 import scalapb.{GeneratedMessage, GeneratedMessageCompanion}
 
 import io.delta.sharing.server.{model, AbfsFileSigner, CausedBy, DeltaSharingIllegalArgumentException, DeltaSharingUnsupportedOperationException, ErrorStrings, GCSFileSigner, PreSignedUrl, S3FileSigner, WasbFileSigner}
-import io.delta.sharing.server.config.{ServerConfig, TableConfig}
+import io.delta.sharing.server.config.TableConfig
 import io.delta.sharing.server.protocol.{QueryTablePageToken, RefreshToken}
 import io.delta.sharing.server.util.JsonUtils
-
-/**
- * A class to load Delta tables from `TableConfig`. It also caches the loaded tables internally
- * to speed up the loading.
- */
-class DeltaSharedTableLoader(serverConfig: ServerConfig) {
-  private val deltaSharedTableCache = {
-    CacheBuilder.newBuilder()
-      .expireAfterAccess(60, TimeUnit.MINUTES)
-      .maximumSize(serverConfig.deltaTableCacheSize)
-      .build[String, DeltaSharedTable]()
-  }
-
-  def loadTable(tableConfig: TableConfig): DeltaSharedTable = {
-    try {
-      val deltaSharedTable =
-        deltaSharedTableCache.get(
-          tableConfig.location,
-          () => {
-            new DeltaSharedTable(
-              tableConfig,
-              serverConfig.preSignedUrlTimeoutSeconds,
-              serverConfig.evaluatePredicateHints,
-              serverConfig.evaluateJsonPredicateHints,
-              serverConfig.evaluateJsonPredicateHintsV2,
-              serverConfig.queryTablePageSizeLimit,
-              serverConfig.queryTablePageTokenTtlMs,
-              serverConfig.refreshTokenTtlMs
-            )
-          }
-        )
-      if (!serverConfig.stalenessAcceptable) {
-        deltaSharedTable.update()
-      }
-      deltaSharedTable
-    } catch {
-      case CausedBy(e: DeltaSharingUnsupportedOperationException) => throw e
-      case e: Throwable => throw e
-    }
-  }
-}
 
 /**
  * A util class stores all query parameters. Used to compute the checksum in the page token for
